demo_streamlit/requirements.txt
streamlit
requests
python-telegram-bot
nest_asyncio
python-dotenv